chore: rename public file to web
